data/Code/BLIP3/requirements.txt
einops
einops-exts
transformers
torch>=2.1.0
pillow
open_clip_torch>=2.16.0 
sentencepiece
